Move handling of negative crossing and strobe values inside MvtxDefs and InttDefs. Fix surface finding from hitsetkey in a few more places.

+++ offline/packages/trackreco/MakeActsGeometry.cc
@@ -1014,7 +1014,7 @@ TrkrDefs::hitsetkey MakeActsGeometry::getInttHitSetKeyFromCoords(unsigned int la
   double check_pos[3] = {0, 0, 0};
   layergeom->find_segment_center(segment_z_bin, segment_phi_bin, check_pos);
 
-  uint16_t crossing = 0;
+  int crossing = 0;
   TrkrDefs::hitsetkey intt_hitsetkey = InttDefs::genHitSetKey(layer, segment_z_bin, segment_phi_bin, crossing);
 
   return intt_hitsetkey;
@@ -1169,7 +1169,7 @@ void MakeActsGeometry::getInttKeyFromNode(TGeoNode *gnode)
   }
 
   // unique key identifying this sensor
-  uint16_t crossing = 0;
+  int crossing = 0;
   TrkrDefs::hitsetkey node_key = InttDefs::genHitSetKey(layer, ladder_z, ladder_phi, crossing);
 
   std::pair<TrkrDefs::hitsetkey, TGeoNode *> tmp = std::make_pair(
